transformations/requirements.txt
paho-mqtt
cassandra-driver
